requirements.txt
google-api-core==2.11.1
google-cloud-pubsub==2.18.1
grgrpcio==1.56.2
grpcio-status==1.56.2
google-auth==2.23.0
smbus==1.1.post2